requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
rich>=13.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
